+++ NEWS.md
@@ -1,3 +1,8 @@
+# acewater NEXT
+
+* Added `autoplot.ace_conditions` and `autoplot.ace_station` to create plots using ggplot2
+
+
 # acewater 0.4.0 (2017-06-09)
 
 * Added stations for all Seattle District projects: Chehalis River Basin, Eastern Washington Rivers, Flathead and Clark Fork Rivers, Green River Basin, Kootenai River Basin, Pend Oreille River Basin, Puyallup River Basin, and Skagit River Basin